corret CMake files to compile fadeColor_sample
